requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
pytest
pytest-mock
psycopg2-binary
aiopg
Pillow
ultralytics
opencv-python-headless
glob2
tqdm
